refactor(gcp): extract Firestore service layer + normalize transaction bytes
